requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
django-allauth==0.62.1
django-ckeditor-5==0.2.13
django-cors-headers==4.3.1
django-filter==24.2
django-js-asset==2.2.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
packaging==24.0
pillow==10.3.0
psycopg2==2.9.9
PyJWT==2.8.0
pyTelegramBotAPI==4.19.1
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
setuptools==70.0.0
sqlparse==0.5.0
telegram==0.0.1
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
